Remove Torso subclass
The Torso subclass is an unnecessary complication that may get in the way of future overhauls, so it has been removed.

Action now has the torso_style property that Torso added, defaulting to None.

Logic based on whether an Action was a Torso now checks if torso_style has a value.