Move torso actions to Torso enumeration
For the sake of readability and organization, all of the actions are going to be moved into enumerations instead of being lumped into a massive list literal.
Until every action is transitioned, enumerations will simply be prepended onto the current list.

Some of the torso comments have been cleaned up or expanded.