Move arms to Appendage enumeration
This continues the transition away from a single action list.

Because limbs and tails are not mechanically distinct, they are not going to get their own enumerations.

Due to the number of members ultimately going in this enumeration, they will be moved piecemeal.

The members of Appendage will be temporarily added to the action list like the other enumerations.