Move joint-adding to Appendage
Adding joints is allowed at the same time adding appendages is, so it belongs in the Appendage enumeration.

In addition to having its constant cost expressed in terms of Value members, adding joints now expresses
the dynamic cost in terms of Warm Amber.

This particular action should be kept above other actions in the enumeration, as it must be performed
before the extra appendages can be added.